requirements.txt
pbr>=0.6,<1.0
SQLAlchemy>=0.7.8,<=0.9.99
amqplib>=0.6.1
anyjson>=0.3.3
argparse
boto>=2.12.0,!=2.13.0
eventlet>=0.13.0
Jinja2
kombu>=2.4.8
lxml>=2.3
Routes>=1.12.3
WebOb>=1.2.3
greenlet>=0.3.2
PasteDeploy>=1.5.0
Paste
sqlalchemy-migrate>=0.8.2,!=0.8.4
netaddr>=0.7.6
suds>=0.4
paramiko>=1.9.0
pyasn1
Babel>=1.3
iso8601>=0.1.9
jsonschema>=2.0.0,<3.0.0
python-cinderclient>=1.0.6
python-neutronclient>=2.3.4,<3
python-glanceclient>=0.9.0
python-keystoneclient>=0.7.0
six>=1.5.2
stevedore>=0.14
websockify>=0.5.1,<0.6
wsgiref>=0.1.2
oslo.config>=1.2.0
oslo.rootwrap
pycadf>=0.4.1
oslo.messaging>=1.3.0a9
websocket_client>=0.24.0
wssh>=0.1.0
httplib2>=0.7.5
